pyproject.toml
[project]
name = "ibots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyntcore>=2026.2.1",
    "tensorflow>=2.20.0",
    "ultralytics>=8.4.5",
    "zombie-imp>=0.0.4",
]
